Fix nondeterministic bug in RecvSelection::deinit
Using Arc::ptr_eq on trait object pointers can fail unpredictably
because of https://github.com/rust-lang/rust/issues/46139.

This can prevent a Hook from being removed when its RecvSelection is
de-inited, which makes it incorrectly push Tokens to a queue owned by a
Selector that no longer exists.

This may be the cause of #44.